First lesson on the code
I read the data into a pandas data frame and then removed a column to clean the data. I then changed all the numbers in all the columns except the 'Unnamed' into the proper format (int64). After that I added an extra column with the total murders of each country. This will essential for the visualisation of my data. Next lesson I will figure out the top 5 countries with the most murders and possibly the top 5 with the least murders. I will also probably visualise the data in a proper chart.