Creating a user interface
I spent the whole lesson creating an easy-to-navigate user interface. I used most of the code from my portfolio and then altered and added some bits. Using the tutorial on google classroom, it was quite easy to create this user interface. However, towards the end, I did run into a few issues but they were simple enough to fix quickly. I am probably going to spend the following lesson on the next part of this assessment.